+++ test/publishing/unit/index.js
@@ -8,6 +8,7 @@ var hostedFields = require('../../../dist/published/hosted-fields');
 var paypal = require('../../../dist/published/paypal');
 var americanExpress = require('../../../dist/published/american-express');
 var unionpay = require('../../../dist/published/unionpay');
+var threeDSecure = require('../../../dist/published/three-d-secure');
 
 describe('braintree', function () {
   it('only exports certain properties', function () {
@@ -18,7 +19,8 @@ describe('braintree', function () {
       'hostedFields',
       'paypal',
       'americanExpress',
-      'unionpay'
+      'unionpay',
+      'threeDSecure'
     ]);
   });
 
@@ -44,6 +46,10 @@ describe('braintree', function () {
 
   it('exports unionpay', function () {
     expect(braintree.unionpay).to.equal(unionpay);
+  });
+
+  it('exports threeDSecure', function () {
+    expect(braintree.threeDSecure).to.equal(threeDSecure);
   });
 
   it('exports VERSION', function () {